refactor(context): dedup finders/MCP tools and wire package-aware resolution
Address the review of the last three days of changes:

- Extract AbstractFinder for the shared depth-bounded BFS traversal, regex
  patterns and comment/literal stripping that Finder and PackageAwareFinder
  duplicated; each now supplies only its direct-resolution step.
- Extract AbstractClassContextTool for the path-confinement, source loading
  and class lookup shared by find_context and estimate_tokens.
- Wire the more precise PackageAwareFinder into both MCP tools so classes
  sharing a simple name across packages are told apart, and default
  estimate_tokens to SubwordTokenEstimator, which tracks real tokenizers more
  closely on source code.
- Document Scala support in the project_tree, find_context and estimate_tokens
  tool schemas and Javadoc.
- Make BudgetedContext.fill iterative to avoid stack growth on large graphs.
- Drop a redundant second comment/literal strip in PackageAwareFinder indexing.
