+++ code/context/src/main/java/io/github/adamw7/context/mcp/EstimateTokensTool.java
@@ -1,24 +1,18 @@
 package io.github.adamw7.context.mcp;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
 import io.github.adamw7.context.ClassContainer;
-import io.github.adamw7.context.Finder;
-import io.github.adamw7.context.HeuristicTokenEstimator;
 import io.github.adamw7.context.Language;
-import io.github.adamw7.context.ProjectSources;
+import io.github.adamw7.context.PackageAwareFinder;
+import io.github.adamw7.context.SubwordTokenEstimator;
 import io.github.adamw7.context.TokenEstimator;
-import io.modelcontextprotocol.spec.McpSchema.CallToolResult;
-import io.modelcontextprotocol.spec.McpSchema.TextContent;
 import io.modelcontextprotocol.spec.McpSchema.Tool;
 
 /**
@@ -29,22 +23,16 @@ import io.modelcontextprotocol.spec.McpSchema.Tool;
  * is reported as an error result rather than an exception so the agent gets a
  * clear, actionable message.
  */
-public class EstimateTokensTool implements ContextTool {
+public class EstimateTokensTool extends AbstractClassContextTool {
 
-	private static final Logger log = LogManager.getLogger(EstimateTokensTool.class.getName());
-
-	private static final int DEFAULT_DEPTH = 1;
-	private static final int MAX_DEPTH = 10;
-
-	private final PathPolicy pathPolicy;
 	private final TokenEstimator estimator;
 
 	public EstimateTokensTool(PathPolicy pathPolicy) {
-		this(pathPolicy, new HeuristicTokenEstimator());
+		this(pathPolicy, new SubwordTokenEstimator());
 	}
 
 	public EstimateTokensTool(PathPolicy pathPolicy, TokenEstimator estimator) {
-		this.pathPolicy = pathPolicy;
+		super(pathPolicy);
 		this.estimator = estimator;
 	}
 
@@ -57,7 +45,7 @@ public class EstimateTokensTool implements ContextTool {
 							"class_name", Map.of("type", "string",
 									"description", "simple name of the class to inspect, e.g. Foo or Foo.java"),
 							"language", Map.of("type", "string",
-									"description", "source language: java (default) or kotlin"),
+									"description", "source language: java (default), kotlin or scala"),
 							"depth", Map.of("type", "integer",
 									"description", "how many levels of transitive dependencies to include (default 1)")),
 					"required", List.of("path", "class_name")))
@@ -70,36 +58,7 @@ public class EstimateTokensTool implements ContextTool {
 	}
 
 	@Override
-	public CallToolResult apply(Map<String, Object> arguments) {
-		log.info("Calling MCP estimate_tokens tool for {}", arguments);
-		Path root = pathPolicy.resolve(ToolArguments.requiredString(arguments, "path"));
-		Language language = ToolArguments.optionalLanguage(arguments, "language", Language.JAVA);
-		int depth = ToolArguments.optionalBoundedInt(arguments, "depth", DEFAULT_DEPTH, 0, MAX_DEPTH);
-		Set<ClassContainer> containers = Set.copyOf(new ProjectSources(language).load(root).values());
-
-		ClassContainer target = findTarget(containers, arguments, language);
-		if (target == null) {
-			return error("Class not found: " + ToolArguments.requiredString(arguments, "class_name"));
-		}
-		return success(estimate(containers, target, language, depth));
-	}
-
-	private ClassContainer findTarget(Set<ClassContainer> containers, Map<String, Object> arguments, Language language) {
-		String fileName = fileNameOf(ToolArguments.requiredString(arguments, "class_name"), language);
-		return containers.stream()
-				.filter(container -> container.className().equals(fileName))
-				.findFirst()
-				.orElse(null);
-	}
-
-	private String fileNameOf(String className, Language language) {
-		if (className.endsWith(language.extension())) {
-			return className;
-		}
-		return className + language.extension();
-	}
-
-	private String estimate(Set<ClassContainer> containers, ClassContainer target, Language language, int depth) {
+	protected String result(Set<ClassContainer> containers, ClassContainer target, Language language, int depth) {
 		List<ClassContainer> context = assembleContext(containers, target, language, depth);
 		JSONArray classes = new JSONArray();
 		int total = 0;
@@ -115,7 +74,7 @@ public class EstimateTokensTool implements ContextTool {
 			Language language, int depth) {
 		List<ClassContainer> context = new ArrayList<>();
 		context.add(target);
-		context.addAll(new Finder(containers, language).find(target, depth));
+		context.addAll(new PackageAwareFinder(containers, language).find(target, depth));
 		return context;
 	}
 
@@ -131,19 +90,5 @@ public class EstimateTokensTool implements ContextTool {
 		report.put("total", total);
 		report.put("classes", classes);
 		return report;
-	}
-
-	private CallToolResult success(String reportJson) {
-		return CallToolResult.builder()
-				.content(List.of(TextContent.builder(reportJson).build()))
-				.isError(false)
-				.build();
-	}
-
-	private CallToolResult error(String message) {
-		return CallToolResult.builder()
-				.content(List.of(TextContent.builder(message).build()))
-				.isError(true)
-				.build();
 	}
 }
